add test for error handling
